64 gedoe

+++ Purplepen/Controllers/UploadController.cs
@@ -52,7 +52,7 @@ namespace Purplepen.Controllers
                     //save into database
                     project p = new project();
                     uploadversion u = new uploadversion();
-                    p.user_id = Convert.ToInt32(Session["fbID"]);
+                    p.user_id = Convert.ToInt64(Session["fbID"]);
                     var blap = form["project_id"];
                     if (form["project_id"] != null)
                     {
